Move presenting results to different class - ResultsPresenter
- add method to SequenceClassiferComparator to get presenter instead of showing results
- with ResultsPresenter you can show all comparisons together
or choose to show e.g. only confusion matrices
